add build_tf_buffer function
